Allow server (universal) rendering
This change allows react-image-lightbox to render on the server side.
Originally, I made some attempts to piggyback on the work at
https://github.com/abhirathore2006/react-image-lightbox-universal
mentioned in #27, but it seemed like using ExtractTextPlugin was adding
some unnecessary complexity.

Instead, this change trades out style-loader for isomorphic-style-loader
and just embeds the CSS within the LightBox component itself.